Add virtual with same entries as aliases
The aliases file is only considered for entries in
"mydestination = $myhostname, localhost.$mydomain, localhost" .
If an email is received for a UNIX account it will be expanded
as username@mydomain. This works well as long as mydomain of the
host is the same as the receiving domain. To allow for proper
resolution of UNIX account target addresses define the
virtual file.